prometheus-grafana/otel/python-app-otlp/gcp-consumer-app/requirements.txt
google-cloud-pubsub==2.21.0
google-cloud-bigquery==3.24.0